Storage Refactor

+++ app/src/main/java/com/arkanzi/udant/feature/archive/manager/ArchiveManager.kt
@@ -3,8 +3,10 @@ package com.arkanzi.udant.feature.archive.manager
 import android.content.Context
 import android.content.Intent
 import androidx.core.content.ContextCompat
+import androidx.core.net.toUri
 import com.arkanzi.udant.core.model.ArchiveStatus
 import com.arkanzi.udant.core.notification.NotificationController
+import com.arkanzi.udant.core.storage.StorageManager
 import com.arkanzi.udant.feature.archive.repository.ArchiveRepository
 import com.arkanzi.udant.feature.archive.service.ArchiveService
 import dagger.hilt.android.qualifiers.ApplicationContext
@@ -20,9 +22,12 @@ class ArchiveManager @Inject constructor(
 
     private val archiveRepository: ArchiveRepository,
 
-    private val notificationController: NotificationController
+    private val notificationController: NotificationController,
+
+    private val storageManager: StorageManager
 
 ) {
+
     suspend fun archive(
         savedArticleId: Long
     ): Result<Unit> {
@@ -84,7 +89,6 @@ class ArchiveManager @Inject constructor(
         notificationController.showArchiveQueued()
 
 
-
         val intent =
             Intent(
                 context,
@@ -113,9 +117,31 @@ class ArchiveManager @Inject constructor(
     suspend fun deleteArchive(
         savedArticleId: Long
     ) {
+        val archiveFolderUri =
+            archiveRepository
+                .getArchiveFolderUri()
+                .firstOrNull()
 
-        archiveRepository.deleteArchive(
-            savedArticleId
-        )
+        val article =
+            archiveRepository
+                .getSavedArticleById(
+                    savedArticleId
+                ) ?: return
+
+        val archiveUri =
+            article.archiveUri
+                ?: return
+
+
+        if (storageManager
+                .deleteFileInSaf(
+                    archiveFolderUri?.toUri(),
+                    archiveUri.toUri()
+                )
+        ) {
+            archiveRepository.deleteArchive(
+                savedArticleId
+            )
+        }
     }
 }
